新增俩接口查询受试者做过的模板列表；根据受试者id查询做过的问卷模板

+++ que-center/src/main/java/com/m3ds/que/center/entity/po/Question.java
@@ -92,6 +92,6 @@ public class Question extends BasePo<Question> {
     /**
      * 序号
      */
-    private long serialNum;
+    private Integer serialNum;
 
 }
